refactor: match to design / support mobile view / make it more reusable
