+++ pages/shorting/index.tsx
@@ -8,7 +8,7 @@ import ShortingHistory from 'sections/shorting/ShortingHistory';
 import ShortingRewards from 'sections/shorting/ShortingRewards';
 
 import AppLayout from 'sections/shared/Layout/AppLayout';
-import { PageContent } from 'styles/common';
+import { GridDiv, PageContent } from 'styles/common';
 import media from 'styles/media';
 import { SYNTHS_MAP } from 'constants/currency';
 import { FlexDivRow } from 'styles/common';
@@ -24,10 +24,10 @@ const Shorting: FC = () => {
 			<AppLayout>
 				<StyledPageContent>
 					<ShortingCard />
-					<FlexDivRow>
+					<ShortingRewardsContainer>
 						<ShortingRewards synth={SYNTHS_MAP.sETH} />
 						<ShortingRewards synth={SYNTHS_MAP.sBTC} />
-					</FlexDivRow>
+					</ShortingRewardsContainer>
 					<ShortingHistory />
 				</StyledPageContent>
 			</AppLayout>
@@ -39,6 +39,16 @@ const StyledPageContent = styled(PageContent)`
 	padding-top: 55px;
 	${media.greaterThan('md')`
 		max-width: 1000px;
+	`}
+`;
+
+const ShortingRewardsContainer = styled(GridDiv)`
+	grid-gap: 24px;
+	grid-auto-flow: column;
+	${media.lessThan('md')`
+		grid-auto-flow: row;
+		/* TODO: this is kinda ugly, and basically undoing the spacing that comes from the TradeSummaryCard */
+		margin-top: -62px;
 	`}
 `;
