+++ sections/exchange/FooterCard/TradeSummaryCard/TradeSummaryCard.tsx
@@ -1,35 +1,32 @@
 import { FC, useMemo } from 'react';
 import { useTranslation, Trans } from 'react-i18next';
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import Tippy from '@tippyjs/react';
-import { customGasPriceState, gasSpeedState } from 'store/wallet';
-import { useRecoilState } from 'recoil';
-import { Svg } from 'react-optimized-image';
 import BigNumber from 'bignumber.js';
 
 import { Synth } from 'lib/synthetix';
 
-import { GasPrices, GAS_SPEEDS } from 'queries/network/useEthGasPriceQuery';
+import { GasPrices } from 'queries/network/useEthGasPriceQuery';
 
-import { NO_VALUE, ESTIMATE_VALUE } from 'constants/placeholder';
+import { NO_VALUE } from 'constants/placeholder';
 import { CurrencyKey } from 'constants/currency';
 
 import { secondsToTime } from 'utils/formatters/date';
 
 import Button from 'components/Button';
 import { DesktopOnlyView, MobileOrTabletView } from 'components/Media';
-import NumericInput from 'components/Input/NumericInput';
 import Card from 'components/Card';
-
-import InfoIcon from 'assets/svg/app/info.svg';
 
 import { formatCurrency, formatPercent } from 'utils/formatters/number';
 
-import { NoTextTransform, numericValueCSS, NumericValue } from 'styles/common';
-import media from 'styles/media';
+import { NoTextTransform } from 'styles/common';
 
 import { MessageContainer, SubmissionDisabledReason } from '../common';
 import useSelectedPriceCurrency from 'hooks/useSelectedPriceCurrency';
+
+import { SummaryItems, SummaryItem, SummaryItemValue, SummaryItemLabel } from '../common';
+
+import GasPriceSummaryItem from './GasPriceSummaryItem';
 
 type TradeSummaryCardProps = {
 	submissionDisabledReason: SubmissionDisabledReason | null;
@@ -73,89 +70,15 @@ const TradeSummaryCard: FC<TradeSummaryCardProps> = ({
 	...rest
 }) => {
 	const { t } = useTranslation();
-	const [gasSpeed, setGasSpeed] = useRecoilState(gasSpeedState);
-	const [customGasPrice, setCustomGasPrice] = useRecoilState(customGasPriceState);
 	const { selectedPriceCurrency } = useSelectedPriceCurrency();
-
-	const hasCustomGasPrice = customGasPrice !== '';
-	const gasPrice = gasPrices ? gasPrices[gasSpeed] : null;
 
 	const isSubmissionDisabled = useMemo(() => (submissionDisabledReason != null ? true : false), [
 		submissionDisabledReason,
 	]);
 
-	const gasPriceItem = hasCustomGasPrice ? (
-		<span data-testid="gas-price">{Number(customGasPrice)}</span>
-	) : (
-		<span data-testid="gas-price">
-			{ESTIMATE_VALUE} {gasPrice}
-		</span>
-	);
-
 	const summaryItems = (
 		<SummaryItems attached={attached}>
-			<SummaryItem>
-				<SummaryItemLabel>{t('exchange.summary-info.gas-price-gwei')}</SummaryItemLabel>
-				<SummaryItemValue>
-					{gasPrice != null ? (
-						<>
-							{transactionFee != null ? (
-								<GasPriceCostTooltip
-									content={
-										<span>
-											{formatCurrency(selectedPriceCurrency.name, transactionFee, {
-												sign: selectedPriceCurrency.sign,
-											})}
-										</span>
-									}
-									arrow={false}
-								>
-									<GasPriceItem>
-										{gasPriceItem}
-										<Svg src={InfoIcon} />
-									</GasPriceItem>
-								</GasPriceCostTooltip>
-							) : (
-								gasPriceItem
-							)}
-							<GasPriceTooltip
-								trigger="click"
-								arrow={false}
-								content={
-									<GasSelectContainer>
-										<CustomGasPriceContainer>
-											<CustomGasPrice
-												value={customGasPrice}
-												onChange={(_, value) => setCustomGasPrice(value)}
-												placeholder={t('common.custom')}
-											/>
-										</CustomGasPriceContainer>
-										{GAS_SPEEDS.map((speed) => (
-											<StyledGasButton
-												key={speed}
-												variant="select"
-												onClick={() => {
-													setCustomGasPrice('');
-													setGasSpeed(speed);
-												}}
-												isActive={hasCustomGasPrice ? false : gasSpeed === speed}
-											>
-												<span>{t(`common.gas-prices.${speed}`)}</span>
-												<NumericValue>{gasPrices![speed]}</NumericValue>
-											</StyledGasButton>
-										))}
-									</GasSelectContainer>
-								}
-								interactive={true}
-							>
-								<StyledGasEditButton role="button">{t('common.edit')}</StyledGasEditButton>
-							</GasPriceTooltip>
-						</>
-					) : (
-						NO_VALUE
-					)}
-				</SummaryItemValue>
-			</SummaryItem>
+			<GasPriceSummaryItem gasPrices={gasPrices} transactionFee={transactionFee} />
 			<SummaryItem>
 				{isCreateShort ? (
 					<>
@@ -249,109 +172,6 @@ const TradeSummaryCard: FC<TradeSummaryCardProps> = ({
 		</>
 	);
 };
-
-export const SummaryItems = styled.div<{ attached?: boolean }>`
-	display: grid;
-	grid-auto-flow: column;
-	flex-grow: 1;
-	${media.lessThan('md')`
-		grid-auto-flow: unset;
-		grid-template-columns: auto auto;
-		grid-template-rows: auto auto;
-		grid-gap: 20px;
-	`}
-
-	${(props) =>
-		props.attached &&
-		css`
-			& {
-				grid-template-rows: unset;
-			}
-		`}
-`;
-
-export const SummaryItem = styled.div`
-	display: grid;
-	grid-gap: 4px;
-	width: 110px;
-	${media.lessThan('md')`
-		width: unset;
-	`}
-`;
-
-export const SummaryItemLabel = styled.div`
-	text-transform: capitalize;
-`;
-
-export const SummaryItemValue = styled.div`
-	color: ${(props) => props.theme.colors.white};
-	${numericValueCSS};
-	max-width: 100px;
-	overflow: hidden;
-	text-overflow: ellipsis;
-`;
-
-export const GasPriceTooltip = styled(Tippy)`
-	background: ${(props) => props.theme.colors.elderberry};
-	border: 0.5px solid ${(props) => props.theme.colors.navy};
-	border-radius: 4px;
-	width: 120px;
-	.tippy-content {
-		padding: 0;
-	}
-`;
-
-export const GasPriceCostTooltip = styled(GasPriceTooltip)`
-	width: auto;
-	font-size: 12px;
-	.tippy-content {
-		padding: 5px;
-		font-family: ${(props) => props.theme.fonts.mono};
-	}
-`;
-
-export const GasSelectContainer = styled.div`
-	padding: 16px 0 8px 0;
-`;
-
-export const CustomGasPriceContainer = styled.div`
-	margin: 0 10px 5px 10px;
-`;
-
-export const CustomGasPrice = styled(NumericInput)`
-	width: 100%;
-	border: 0;
-	font-size: 12px;
-	::placeholder {
-		font-family: ${(props) => props.theme.fonts.mono};
-	}
-`;
-
-export const StyledGasButton = styled(Button)`
-	width: 100%;
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	padding-left: 10px;
-	padding-right: 10px;
-`;
-
-export const GasPriceItem = styled.span`
-	display: inline-flex;
-	align-items: center;
-	cursor: pointer;
-	svg {
-		margin-left: 5px;
-	}
-`;
-
-export const StyledGasEditButton = styled.span`
-	font-family: ${(props) => props.theme.fonts.bold};
-	padding-left: 5px;
-	cursor: pointer;
-	color: ${(props) => props.theme.colors.goldColors.color3};
-	text-transform: uppercase;
-`;
 
 export const ErrorTooltip = styled(Tippy)`
 	font-size: 12px;
