+++ sections/shorting/ShortingRewards/ShortingRewards.tsx
@@ -5,47 +5,40 @@ import { useRecoilState, useRecoilValue } from 'recoil';
 import { Svg } from 'react-optimized-image';
 import { ethers } from 'ethers';
 
-import { DesktopOnlyView, MobileOrTabletView } from 'components/Media';
-import {} from 'styles/common';
-import Card from 'components/Card';
+import Notify from 'containers/Notify';
+import Connector from 'containers/Connector';
+
 import synthetix from 'lib/synthetix';
 import Button from 'components/Button';
 import { normalizeGasLimit, getTransactionPrice, gasPriceInWei } from 'utils/network';
 import { getExchangeRatesForCurrencies } from 'utils/currencies';
-import InfoIcon from 'assets/svg/app/info.svg';
 import TxConfirmationModal from 'sections/shared/modals/TxConfirmationModal';
 
 import ArrowRightIcon from 'assets/svg/app/circle-arrow-right.svg';
-import { GAS_SPEEDS } from 'queries/network/useEthGasPriceQuery';
-import { FixedFooterMixin, GridDivCentered, NumericValue } from 'styles/common';
+
+import { FixedFooterMixin, GridDivCentered } from 'styles/common';
 import media from 'styles/media';
-import { gasSpeedState, customGasPriceState, walletAddressState } from 'store/wallet';
+
+import { gasSpeedState, walletAddressState } from 'store/wallet';
 import useSelectedPriceCurrency from 'hooks/useSelectedPriceCurrency';
-import { formatCurrency, formatCryptoCurrency } from 'utils/formatters/number';
+import { formatCryptoCurrency } from 'utils/formatters/number';
+
 import useEthGasPriceQuery from 'queries/network/useEthGasPriceQuery';
-import { SYNTHS_MAP } from 'constants/currency';
 import useExchangeRatesQuery from 'queries/rates/useExchangeRatesQuery';
 import useCollateralShortDataQuery from 'queries/collateral/useCollateralShortDataQuery';
+
+import { SYNTHS_MAP } from 'constants/currency';
 import { CurrencyKey } from 'constants/currency';
-import { NO_VALUE, ESTIMATE_VALUE } from 'constants/placeholder';
-import Notify from 'containers/Notify';
-import Connector from 'containers/Connector';
+
+import GasPriceSummaryItem from 'sections/exchange/FooterCard/TradeSummaryCard/GasPriceSummaryItem';
 
 import {
 	SummaryItems,
 	SummaryItem,
 	SummaryItemLabel,
 	SummaryItemValue,
-	GasPriceCostTooltip,
-	GasPriceItem,
-	GasPriceTooltip,
-	StyledGasButton,
-	CustomGasPrice,
-	GasSelectContainer,
-	CustomGasPriceContainer,
-	MobileCard,
-	StyledGasEditButton,
-} from 'sections/exchange/FooterCard/TradeSummaryCard';
+} from 'sections/exchange/FooterCard/common';
+
 import { SubmissionDisabledReason } from 'sections/exchange/FooterCard/common';
 
 interface ShortingRewardsProps {
@@ -54,8 +47,7 @@ interface ShortingRewardsProps {
 
 const ShortingRewards: FC<ShortingRewardsProps> = ({ synth }) => {
 	const { t } = useTranslation();
-	const [gasSpeed, setGasSpeed] = useRecoilState(gasSpeedState);
-	const [customGasPrice, setCustomGasPrice] = useRecoilState(customGasPriceState);
+	const [gasSpeed] = useRecoilState(gasSpeedState);
 	const walletAddress = useRecoilValue(walletAddressState);
 	const { selectedPriceCurrency } = useSelectedPriceCurrency();
 	const [txConfirmationModalOpen, setTxConfirmationModalOpen] = useState<boolean>(false);
@@ -110,9 +102,8 @@ const ShortingRewards: FC<ShortingRewardsProps> = ({ synth }) => {
 		getGasEstimateCall();
 	}, [getGasEstimate]);
 
-	const gasPrices = useMemo(() => ethGasPriceQuery?.data ?? null, [ethGasPriceQuery.data]);
+	const gasPrices = useMemo(() => ethGasPriceQuery?.data ?? undefined, [ethGasPriceQuery.data]);
 
-	const hasCustomGasPrice = customGasPrice !== '';
 	const gasPrice = ethGasPriceQuery?.data != null ? ethGasPriceQuery.data[gasSpeed] : null;
 	const exchangeRates = exchangeRatesQuery.isSuccess ? exchangeRatesQuery.data ?? null : null;
 
@@ -126,14 +117,6 @@ const ShortingRewards: FC<ShortingRewardsProps> = ({ synth }) => {
 		gasLimit,
 		ethPriceRate,
 	]);
-
-	const gasPriceItem = hasCustomGasPrice ? (
-		<span data-testid="gas-price">{Number(customGasPrice)}</span>
-	) : (
-		<span data-testid="gas-price">
-			{ESTIMATE_VALUE} {gasPrice}
-		</span>
-	);
 
 	const onSubmit = async () => {
 		if (synthetix.js != null && gasPrice != null) {
@@ -181,85 +164,19 @@ const ShortingRewards: FC<ShortingRewardsProps> = ({ synth }) => {
 	const summaryItems = (
 		<SummaryItems attached={false}>
 			<SummaryItem>
-				<SummaryItemLabel>{t('shorting.rewards.available')}</SummaryItemLabel>
-				<SummaryItemValue>
+				<StyledSummaryItemLabel>{t('shorting.rewards.available')}</StyledSummaryItemLabel>
+				<BoldSummaryItemValue>
 					{formatCryptoCurrency(shortingRewards ?? 0, { currencyKey: synth })}
-				</SummaryItemValue>
+				</BoldSummaryItemValue>
 			</SummaryItem>
-			<SummaryItem>
-				<SummaryItemLabel>{t('exchange.summary-info.gas-price-gwei')}</SummaryItemLabel>
-				<SummaryItemValue>
-					{gasPrice != null ? (
-						<>
-							{transactionFee != null ? (
-								<GasPriceCostTooltip
-									content={
-										<span>
-											{formatCurrency(selectedPriceCurrency.name, transactionFee, {
-												sign: selectedPriceCurrency.sign,
-											})}
-										</span>
-									}
-									arrow={false}
-								>
-									<GasPriceItem>
-										{gasPriceItem}
-										<Svg src={InfoIcon} />
-									</GasPriceItem>
-								</GasPriceCostTooltip>
-							) : (
-								gasPriceItem
-							)}
-							<GasPriceTooltip
-								trigger="click"
-								arrow={false}
-								content={
-									<GasSelectContainer>
-										<CustomGasPriceContainer>
-											<CustomGasPrice
-												value={customGasPrice}
-												onChange={(_, value) => setCustomGasPrice(value)}
-												placeholder={t('common.custom')}
-											/>
-										</CustomGasPriceContainer>
-										{GAS_SPEEDS.map((speed) => (
-											<StyledGasButton
-												key={speed}
-												variant="select"
-												onClick={() => {
-													setCustomGasPrice('');
-													setGasSpeed(speed);
-												}}
-												isActive={hasCustomGasPrice ? false : gasSpeed === speed}
-											>
-												<span>{t(`common.gas-prices.${speed}`)}</span>
-												<NumericValue>{gasPrices![speed]}</NumericValue>
-											</StyledGasButton>
-										))}
-									</GasSelectContainer>
-								}
-								interactive={true}
-							>
-								<StyledGasEditButton role="button">{t('common.edit')}</StyledGasEditButton>
-							</GasPriceTooltip>
-						</>
-					) : (
-						NO_VALUE
-					)}
-				</SummaryItemValue>
-			</SummaryItem>
+			<GasPriceSummaryItem gasPrices={gasPrices} transactionFee={transactionFee} />
 		</SummaryItems>
 	);
 
 	return (
 		<>
-			<MobileOrTabletView>
-				<MobileCard className="trade-summary-card">
-					<Card.Body>{summaryItems}</Card.Body>
-				</MobileCard>
-			</MobileOrTabletView>
 			<MessageContainer attached={false} className="footer-card">
-				<DesktopOnlyView>{summaryItems}</DesktopOnlyView>
+				{summaryItems}
 				<Button
 					variant="primary"
 					isRounded={true}
@@ -295,22 +212,22 @@ const ShortingRewards: FC<ShortingRewardsProps> = ({ synth }) => {
 };
 
 export const MessageContainer = styled(GridDivCentered)<{ attached?: boolean }>`
-	width: 48%;
 	border-radius: 4px;
 	grid-template-columns: 1fr auto;
-	background-color: ${(props) => props.theme.colors.elderberry};
+	background-color: ${(props) => props.theme.colors.navy};
 	padding: 16px 32px;
-	${(props) =>
-		props.attached &&
-		css`
-			border-radius: 4px;
-		`}
 	${media.lessThan('md')`
-		${FixedFooterMixin};
-		box-shadow: 0 -8px 8px 0 ${(props) => props.theme.colors.black};
-		justify-content: center;
-		display: flex;
+		grid-template-columns: unset;
+		grid-template-rows: 1fr auto;
 	`}
+`;
+
+const StyledSummaryItemLabel = styled(SummaryItemLabel)`
+	color: ${(props) => props.theme.colors.silver};
+`;
+
+const BoldSummaryItemValue = styled(SummaryItemValue)`
+	font-family: ${(props) => props.theme.fonts.bold};
 `;
 
 export default ShortingRewards;
